requirements_demo.txt
# Core ML/AI libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.20.0

# RLHF specific
peft>=0.4.0
trl>=0.7.0

# Evaluation metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Demo interface
gradio>=4.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
